refactor: rename incident module
