+++ src/components/probe/prober/http/index.ts
@@ -36,7 +36,7 @@ import responseChecker from '../../../../plugins/validate-response/checkers'
 import { getAlertID } from '../../../../utils/alert-id'
 import { getEventEmitter } from '../../../../utils/events'
 import { isSymonModeFrom } from '../../../config'
-import { addIncident } from '../../../downtime-counter'
+import { addIncident } from '../../../incident'
 import { saveProbeRequestLog } from '../../../logger/history'
 import { logResponseTime } from '../../../logger/response-time-log'
 import { httpRequest } from './request'
