+++ src/symon/index.test.ts
@@ -34,7 +34,7 @@ import SymonClient from '.'
 import { getContext, resetContext, setContext } from '../context'
 import { deleteProbe, findProbe, getProbes } from '../components/config/probe'
 import { validateProbes } from '../components/config/validation'
-import { addIncident, findIncident } from '../components/downtime-counter'
+import { addIncident, findIncident } from '../components/incident'
 import events from '../events'
 import { getEventEmitter } from '../utils/events'
 import { getErrorMessage } from '../utils/catch-error-handler'
